# backend/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CollabMate API",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/task": {
      "post": {
        "summary": "Create a new task",
        "tags": [
          "Task"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "userId"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Task created"
          },
          "400": {
            "description": "Missing title or userId"
          },
          "401": {
            "description": "Unauthorized - missing or invalid token"
          }
        }
      },
      "get": {
        "summary": "Get tasks",
        "description": "Optionally filtered by status, teamId, userId",
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "teamId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filtered task list"
          }
        }
      }
    },
    "/task/{taskId}": {
      "put": {
        "summary": "Update a task by ID",
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Task updated"
          },
          "401": {
            "description": "Unauthorized - missing or invalid token"
          }
        }
      },
      "delete": {
        "summary": "Delete a task by ID (with cascade update user references)",
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "task": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/task/{taskId}/assign": {
      "patch": {
        "summary": "Assign a user to a task",
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User assigned to task"
          },
          "401": {
            "description": "Unauthorized - missing or invalid token"
          }
        }
      }
    },
    "/task/{taskId}/status": {
      "patch": {
        "summary": "Change task status",
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "newStatus"
                ],
                "properties": {
                  "newStatus": {
                    "type": "string",
                    "enum": [
                      "To Do",
                      "In Progress",
                      "Done"
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Task status updated"
          },
          "401": {
            "description": "Unauthorized - missing or invalid token"
          }
        }
      }
    },
    "/task/grouped": {
      "get": {
        "summary": "Get tasks grouped by status or team",
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "by",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "status",
                "team"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Grouped tasks object"
          },
          "400": {
            "description": "Invalid grouping method"
          }
        }
      }
    },
    "/task/all": {
      "delete": {
        "summary": "Delete all tasks (cascade update user.taskId/tasks)",
        "description": "Deletes all tasks in the system.\nAlso clears all users' `taskId` and `tasks` fields.\n",
        "tags": [
          "Task"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All tasks deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}